Fix dependency injection: use localised packages
So far we injected the local workspace when building the unpacked package.
However, when we are in a workspace with changes compared to dependencies
that are already released, cargo sometimes pulls in the remote and local
package in the check then has dependency mismatches.

This change fixes that by collecting the locally build packages during the
checking-process, and injects only those that it previously build. This
prevents the issue by not ever introducing packages from the local workspace.